[FEAT] 유저 / 배송기사 서비스에서 수정/ 삭제 시 검증 로직 작성 (#150)
* feat: user 이메일 validate 추가

* refactor: email -> id 검증 변환

* feat: user 검증 로직 구현

* feat: courier 검증 로직 구현

* feat: courier 검증 로직 구현 #2

* refactor: CourierService 기능 수정

* refactor: CourierServiceTest 변경

* refactor: UserServiceTest 변경

* refactor: User & Courier 관련 통합테스트 수정

* refactor: User Service 검증 로직 수정

* refactor: 머지 후 충돌 수정

* refactor: Preconditions final 선언

* refactor: UserServiceTest 수정

* refactor: UserService 테스트코드 수정

* refactor: UserService 테스트코드 수정

* refactor: CourierService 테스트코드 수정

+++ src/main/java/com/kt/service/UserServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.kt.constant.UserRole;
+import com.kt.constant.message.ErrorCode;
 import com.kt.domain.dto.request.UserRequest;
 import com.kt.domain.dto.request.SignupRequest;
 import com.kt.domain.dto.response.OrderProductResponse;
@@ -19,8 +20,8 @@ import com.kt.domain.entity.OrderEntity;
 import com.kt.domain.entity.UserEntity;
 import com.kt.repository.orderproduct.OrderProductRepository;
 import com.kt.repository.OrderRepository;
-import com.kt.repository.review.ReviewRepository;
 import com.kt.repository.user.UserRepository;
+import com.kt.util.Preconditions;
 
 import lombok.RequiredArgsConstructor;
 
@@ -33,12 +34,14 @@ public class UserServiceImpl implements UserService {
 	private final OrderRepository orderRepository;
 	private final UserRepository userRepository;
 	private final PasswordEncoder passwordEncoder;
-	private final ReviewRepository reviewRepository;
+
 
 	@Override
-	public UserResponse.Orders getOrdersByUserId(UUID id) {
-		List<OrderEntity> orders = orderRepository.findAllByOrderBy_Id(id);
-		return UserResponse.Orders.of(id, orders);
+	public UserResponse.Orders getOrdersByUserId(UUID currentId, UUID subjectId) {
+		checkReadPermission(currentId, subjectId);
+
+		List<OrderEntity> orders = orderRepository.findAllByOrderBy_Id(subjectId);
+		return UserResponse.Orders.of(subjectId, orders);
 	}
 
 	@Override
@@ -47,13 +50,17 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public Page<UserResponse.Search> getUsers(Pageable pageable, String keyword, UserRole role) {
+	public Page<UserResponse.Search> getUsers(UUID userId, Pageable pageable, String keyword, UserRole role) {
+		checkAdmin(userId);
+
 		return userRepository.searchUsers(pageable, keyword, role);
 	}
 
 	@Override
-	public UserResponse.UserDetail getUserDetail(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getUserDetail(UUID currentId, UUID subjectId) {
+		checkReadPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		return new UserResponse.UserDetail(
 			user.getId(),
 			user.getName(),
@@ -66,8 +73,8 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public UserResponse.UserDetail getAdminDetail(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getUserDetailSelf(UUID currentId) {
+		UserEntity user = userRepository.findByIdOrThrow(currentId);
 		return new UserResponse.UserDetail(
 			user.getId(),
 			user.getName(),
@@ -80,31 +87,57 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void enableUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getAdminDetail(UUID currentId, UUID subjectId) {
+		checkReadPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
+		return new UserResponse.UserDetail(
+			user.getId(),
+			user.getName(),
+			user.getEmail(),
+			user.getRole(),
+			user.getGender(),
+			user.getBirth(),
+			user.getMobile()
+		);
+	}
+
+	@Override
+	public void enableUser(UUID currentId, UUID subjectId) {
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.enabled();
 	}
 
 	@Override
-	public void disableUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void disableUser(UUID currentId, UUID subjectId) {
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.disabled();
 	}
 
 	@Override
-	public void deleteUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void deleteUser(UUID currentId, UUID subjectId) {
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.delete();
 	}
 
 	@Override
-	public void retireUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void retireUser(UUID currentId, UUID subjectId) {
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.retired();
 	}
 
 	@Override
-	public void createAdmin(SignupRequest.SignupMember request) {
+	public void createAdmin(UUID userId, SignupRequest.SignupMember request) {
+		checkAdmin(userId);
+
 		UserEntity admin = UserEntity.create(
 			request.name(),
 			request.email(),
@@ -118,25 +151,34 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void deleteUserPermanently(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void deleteUserPermanently(UUID currentId, UUID subjectId) {
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		orderRepository.clearUser(user.getId());
 		userRepository.delete(user);
 	}
 
 	@Override
-	public void deleteAdmin(UUID adminId) {
+	public void deleteAdmin(UUID currentId, UUID adminId) {
+		checkAdmin(adminId);
+		checkModifyPermission(currentId, adminId);
+
 		UserEntity user = userRepository.findByIdOrThrow(adminId);
 		user.delete();
 	}
 
 	@Override
 	public void updateUserDetail(
-		UUID userId,
+		UUID currentId,
+		UUID subjectId,
 		UserRequest.UpdateDetails details
 	) {
-		UserEntity user = userRepository.findByIdOrThrow(userId);
-		user.updateDetails(
+		checkModifyPermission(currentId, subjectId);
+
+		UserEntity subjectUser = userRepository.findByIdOrThrow(subjectId);
+
+		subjectUser.updateDetails(
 			details.name(),
 			details.mobile(),
 			details.birth(),
@@ -144,4 +186,50 @@ public class UserServiceImpl implements UserService {
 		);
 	}
 
+	@Override
+	public void updateUserDetailSelf(
+		UUID userId,
+		UserRequest.UpdateDetails details
+	) {
+		UserEntity subjectUser = userRepository.findByIdOrThrow(userId);
+
+		subjectUser.updateDetails(
+			details.name(),
+			details.mobile(),
+			details.birth(),
+			details.gender()
+		);
+	}
+
+	private void checkAdmin(UUID currentUserId) {
+		UserEntity user = userRepository.findByIdOrThrow(currentUserId);
+		Preconditions.validate(
+			user.getRole() == UserRole.ADMIN,
+			ErrorCode.NOT_ADMIN
+		);
+	}
+
+	private void checkReadPermission(UUID currentId, UUID subjectId){
+		UserEntity currentUser = userRepository.findByIdOrThrow(currentId);
+		Preconditions.validate(
+			currentUser.getRole().equals(UserRole.ADMIN) | currentId.equals(subjectId),
+			ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
+		);
+	}
+
+	private void checkModifyPermission(UUID currentId, UUID subjectId) {
+		UserEntity subjectUser = userRepository.findByIdOrThrow(subjectId);
+		if (subjectUser.getRole() == UserRole.ADMIN) {
+			Preconditions.validate(
+				currentId.equals(subjectId),
+				ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
+			);
+		} else {
+			UserEntity currentUser = userRepository.findByIdOrThrow(currentId);
+			Preconditions.validate(
+				currentUser.getRole().equals(UserRole.ADMIN) | currentId.equals(subjectId),
+				ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
+			);
+		}
+	}
 }
